Done with this Project

+++ controllers/contacts.js
@@ -3,29 +3,34 @@ import {Contact} from "../models/contact.js";
 
 //@desc: Get all contacts
 //@route: GET /api/contacts
-//@access: Public
+//@access: Private
 export const allContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id});
     res.send(contacts);
 });
 
 //@desc: Get one contact
 //@route: GET /api/contacts/:id
-//@access: Public
+//@access: Private
 export const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact){
         res.status(404);
         throw new Error("Contact not Found");
     }
+
+    if (contact.user_id !== req.user.id){
+        res.status(403);
+        throw new Error("User doesn't have permission to update other user contacts");
+    }
+    
     res.status(200).send(contact);
 });
 
 //@desc: Create new contact
 //@route: POST /api/contacts/
-//@access: Public
+//@access: Private
 export const createContact = asyncHandler(async (req, res) => {
-    console.log("The Request Body is:", req.body);
     const {name, email, phone} = req.body;
     if (!name || !email || !phone) {
         res.status(400);
@@ -34,19 +39,25 @@ export const createContact = asyncHandler(async (req, res) => {
     const contact = await Contact.create({
         name, 
         email,
-        phone
+        phone,
+        user_id: req.user.id
     });
     res.send(contact);
 });
 
 //@desc: Edit a contact
 //@route: PUT /api/contacts/:id
-//@access: Public
+//@access: Private
 export const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact){
         res.status(404);
         throw new Error("Contact not Found");
+    }
+
+    if (contact.user_id !== req.user.id){
+        res.status(403);
+        throw new Error("User doesn't have permission to update other user contacts");
     }
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
@@ -58,13 +69,19 @@ export const updateContact = asyncHandler(async (req, res) => {
 
 //@desc: Delete a contact
 //@route: DELETE /api/contacts/:id
-//@access: Public
+//@access: Private
 export const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact){
         res.status(404);
         throw new Error("Contact not Found");
     }
+
+    if (contact.user_id !== req.user.id){
+        res.status(403);
+        throw new Error("User doesn't have permission to update other user contacts");
+    }
+
     await Contact.findByIdAndRemove(req.params.id);
     res.status(200).send(contact);
 });
